Fix: Исправил MpaController и MpaServiceImpl для использования сервисного слоя
