+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -5,9 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.yandex.practicum.filmorate.dao.MpaDao;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Mpa;
+import ru.yandex.practicum.filmorate.service.MpaService;
 
 import java.util.List;
 
@@ -15,16 +14,15 @@ import java.util.List;
 @RequestMapping("/mpa")
 @RequiredArgsConstructor
 public class MpaController {
-    private final MpaDao mpaDao;
+    private final MpaService mpaService;
 
     @GetMapping
     public List<Mpa> getAllMpa() {
-        return mpaDao.getAllMpa();
+        return mpaService.getAllMpa();
     }
 
     @GetMapping("/{id}")
     public Mpa getMpaById(@PathVariable Long id) {
-        return mpaDao.getMpaById(id)
-                .orElseThrow(() -> new NotFoundException("Рейтинг MPA с ID " + id + " не найден"));
+        return mpaService.getMpaById(id);
     }
 }
